requirements.txt
-f /usr/share/pip-wheels
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-google-analytics-app==5.0.2
django-js-asset==1.2.2
django-robots==4.0
idna==2.10
kombu==4.6.11
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
urllib3==1.26.4
vine==1.3.0
